// csm.web.app/src/app/shared/jobs/api/enums/job-status.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2022
 *
 * *************************************************************************
 */

export enum JobStatusEnum {
    Queued = 'QUEUED',
    Running = 'RUNNING',
    Completed = 'COMPLETED',
    Failed = 'FAILED',
    Rejected = 'REJECTED',
}