// csm.web.app/src/app/shared/misc/enums/file.type.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2022
 *
 * *************************************************************************
 */

export enum FileTypeEnum {
    PDF = 'PDF',
    CSV = 'CSV',
    JSON = 'JSON',
}